4.21*

+++ controllers/blogs.js
@@ -29,6 +29,7 @@ blogsRouter.post('/', async (request, response) => {
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token invalid' })
   }
+
   const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
@@ -54,6 +55,18 @@ blogsRouter.delete('/all', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if(blog.user.toString() !== decodedToken.id.toString()){
+    return response.status(403).json({ error: 'you can only delete blogs you created' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
